chore: update config for about page
